pyproject.toml
[project]
name = "rag-chatbot"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4"
dependencies = [
    "langchain-text-splitters (>=0.3.8,<0.4.0)",
    "langchain-community (>=0.3.24,<0.4.0)",
    "langgraph (>=0.4.5,<0.5.0)",
    "langchain (>=0.3.25,<0.4.0)",
    "langchain-openai (>=0.3.17,<0.4.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "langchain-core (>=0.3.60,<0.4.0)",
    "pymupdf (>=1.25.5,<2.0.0)",
    "langchain-chroma (>=0.2.4,<0.3.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "unstructured (>=0.17.2,<0.18.0)",
    "flask (>=3.1.1,<4.0.0)",
    "flask-socketio (>=5.5.1,<6.0.0)",
    "eventlet (>=0.40.0,<0.41.0)"
]

[tool.poetry]
packages = [{include = "rag_chatbot", from = "src"}]


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
